컴포트 분리 및 수정
